chore: refactor useRequiredChain hook location
