+++ apps/frontend/src/app/3_organisms/EmailNotificationSettingsDialog/EmailNotificationSettingsDialog.tsx
@@ -20,8 +20,8 @@ import {
 import { RSK_CHAIN_ID } from '../../../config/chains';
 
 import { NetworkBanner } from '../../2_molecules/NetworkBanner/NetworkBanner';
-import { useRequiredChain } from '../../2_molecules/NetworkBanner/hooks/useRequiredChain';
 import { useNotificationContext } from '../../../contexts/NotificationContext';
+import { useRequiredChain } from '../../../hooks/chain/useRequiredChain';
 import { useAccount } from '../../../hooks/useAccount';
 import { useChainStore } from '../../../hooks/useChainStore';
 import { translations } from '../../../locales/i18n';
